refactor(frontend): rename state/layers/view-layers
The how-to guide, and `nismod/irv-frontend`, use `state/layers/data-layers/` for map data layers.
