+++ frontend/src/state/layers/data-layers/droughtOptions.ts
